requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorlog==6.8.2
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.1.2
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-timezone-field==7.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
plotly==5.24.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
